pyproject.toml
[tool.poetry]
name = "test-work"
version = "0.1.0"
description = ""
authors = ["vladimirmac <vladimirbodrin89@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.7"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.3.0"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
httpx = "^0.25.2"
aiofiles = "^23.2.1"
pillow = "^10.1.0"
celery = {extras = ["redis"], version = "^5.3.6"}
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
